+++ src/cachelib/serializers.py
@@ -101,3 +101,14 @@ class RedisSerializer(BaseSerializer):
         except ValueError:
             # before 0.8 we did not have serialization. Still support that.
             return value
+
+
+class DynamoDbSerializer(RedisSerializer):
+    """Default serializer for DynamoDbCache."""
+
+    def loads(self, value: _t.Any) -> _t.Any:
+        """The reversal of :meth:`dump_object`. This might be called with
+        None.
+        """
+        value = value.value
+        return super().loads(value)