+++ CHANGES.rst
@@ -4,7 +4,7 @@ Version 0.10.0
 Unreleased
 
 - Improve error message when ``FileSystemCache`` methods are called with non-str keys. :pr:`170`
-
+- Added ``DynamoDb`` as a cache backend
 
 Version 0.9.0
 -------------